major updates: home, trip, posts
major updates: home, trip, posts

+++ app/Http/Controllers/Post/PostController.php
@@ -12,6 +12,8 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use Geokit;
+use App\Setting;
 
 class PostController extends Controller
 {
@@ -50,6 +52,7 @@ class PostController extends Controller
         $this->validate($request, [
             'name' => 'required|max:255',
             'description' => 'required|min:10',
+            'departure_date' => 'required|date',
             'departure_pcode' => 'required',
             'destination_pcode' => 'required',
             'num_riders' => 'required|numeric',
@@ -81,10 +84,20 @@ class PostController extends Controller
         ]));
 
         //Setting other post attributes.
+        $post->departure_date = $request->departure_date;
         $post->departure_pcode = $departure_pcode['postal_code'];
         $post->destination_pcode = $destination_pcode['postal_code'];
         $post->poster_id = Auth::user()->id;
-        $post->cost = 90; //TODO: Determine cost of trip per kilometer.
+
+
+        // Determine the cost based on distance
+        $math = new Geokit\Math();
+        $distance = $math->distanceHaversine([$departure_pcode['latitude'], $departure_pcode['longitude']],
+            [$destination_pcode['latitude'], $destination_pcode['longitude']])
+            ->kilometers();
+        $cost = Setting::find('ride_initial_cost')->value + ($distance * Setting::find('fuel_cost_per_kilometer')->value);
+        $post->cost = round($cost,2);
+
         $post->is_request = false; //TODO: Is request of trip.
 
         $trip = null;
@@ -100,32 +113,40 @@ class PostController extends Controller
                 'every_thur' => 'required_if:one_time,false|boolean',
                 'every_fri' => 'required_if:one_time,false|boolean',
                 'every_sat' => 'required_if:one_time,false|boolean',
-                'time' => 'required_if:one_time,true',
-                'depature_date' => 'required_if:one_time,true:date'
+                'time' => 'required|date_format:H:i',
+                'departure_date' => 'required_if:one_time,true:date'
             ]);
 
             $frequency = $request->only(['every_sun', 'every_mon', 'every_tues', 'every_wed', 'every_thur', 'every_fri', 'every_sat']);
             $frequency = array_values($frequency); //Returns array of the values only
+            $atLeastOne  = false;
+            for($i = 0; $i < count($frequency); $i++) {
+                if($frequency[$i]) {
+                    $atLeastOne = true;
+                }
+            }
+            if(!$atLeastOne) {
+                return back()->withErrors(['frequency' => 'At least one day of the week must be selected.']);
+            }
 
             $trip = new LocalTrip([
                 'frequency' => $frequency,
                 'departure_time' => (new Carbon($request->time))->toTimeString()
             ]);
 
-            $post->departure_date = $request->departure_date;
-
         } else {
             //If post is type of long distance
 
             $this->validate($request, [
-                //TODO incomplete
+                'frequency'=>'required|min:0|max:8'
             ]);
 
             $trip = new LongDistanceTrip([
                 'departure_city' => $departure_pcode['city'],
                 'departure_province' => $departure_pcode['province'],
                 'destination_city' => $destination_pcode['city'],
-                'destination_province' => $destination_pcode['province']
+                'destination_province' => $destination_pcode['province'],
+                'frequency' => $request->frequency
             ]);
 
         }
@@ -181,12 +202,51 @@ class PostController extends Controller
     {
         $post = Post::findOrFail($id);
 
-        if (! $this->canEdit()){
+        if (! $this->canEdit()) {
             return abort(403);
         }
 
+        // Avoiding updating the postal codes for nothing
+        if($post->departure_pcode != $request->departure_pcode
+            || $post->destination_pcode != $request->destination_pcode) {
+
+            // Obtain the new postal coder
+            $coder = new PostalCoder();
+            $departure_pcode = $coder->geocode($request->departure_pcode);
+            if (! $departure_pcode){
+                return back()->withErrors([
+                    'departure_pcode' => 'Invalid Postal Code.'
+                ]);
+            }
+            $departure_pcode = $departure_pcode->toArray();
+
+            $destination_pcode = $coder->geocode($request->destination_pcode);
+            if (! $destination_pcode){
+                return back()->withErrors([
+                    'destination_pcode' => 'Invalid Postal Code.'
+                ]);
+            }
+            $destination_pcode = $destination_pcode->toArray();
+
+            // Update the postal code
+            $post->departure_pcode = $departure_pcode['postal_code'];
+            $post->destination_pcode = $destination_pcode['postal_code'];
+
+            // Update the cost based on (new) distance
+            $math = new Geokit\Math();
+            $distance = $math->distanceHaversine([$departure_pcode['latitude'], $departure_pcode['longitude']],
+                [$destination_pcode['latitude'], $destination_pcode['longitude']])
+                ->kilometers();
+            $cost = Setting::find('ride_initial_cost')->value + ($distance * Setting::find('fuel_cost_per_kilometer')->value);
+            $post->cost = round($cost,2);
+        }
+
+        //Setting other post attributes.
+        $post->departure_date = $request->departure_date;
+
+        $post->save();
         return redirect(route('post.show', ['post' => $post]))
-            ->with('success', 'Your post is now live!');
+            ->with('success', 'Your post has been updated successfully!');
     }
 
     /**
@@ -198,7 +258,11 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
-        $post->delete();
+        if(Auth::user()->id == $post->poster_id) {
+            $post->delete();
+        } else {
+            // Does not have permission to delete this post
+        }
 
         return redirect(route('user.show', ['user' => Auth::user()->id]))
             ->with('success', 'Your post has been deleted!');
